+++ actix-web-lab/src/extract.rs
@@ -3,6 +3,7 @@
 /// An alias for [`actix_web::web::Data<T>`] with a more descriptive name.
 pub type SharedData<T> = actix_web::web::Data<T>;
 
+#[allow(deprecated)]
 pub use crate::body_hash::BodyHash;
 pub use crate::body_hmac::{BodyHmac, HmacConfig};
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
@@ -10,8 +11,6 @@ pub use crate::lazy_data::LazyData;
 pub use crate::local_data::LocalData;
 pub use crate::path::Path;
 pub use crate::query::Query;
-#[doc(hidden)]
-pub use crate::request_hash::{RequestHash, RequestHasher};
 #[doc(hidden)]
 pub use crate::request_signature::{
     RequestSignature, RequestSignatureError, RequestSignatureScheme,